docs: split external and internal setup
For environment and test setup, it makes sense to split these into
internal and external documentations. The reason is that the `make`
commands in the `CONTRIBUTING.md` were heavily reliant on our own
infrastructure. Now, external developers should be able to get up and
running without bazel and LDAP credentials.

In addition, sanity checks were added to make sure that we don't break
the external contribution guide.